requirements.txt
streamlit==1.19.0
scikit-learn==1.2.1
numpy==1.23.5
pandas==1.5.3
pillow==9.3.0
joblib==1.2.0
tensorflow==2.10.0
